Activity

+++ api/src/features/wallets/wallets.resolver.ts
@@ -9,12 +9,13 @@ import {
   Resolver,
 } from '@nestjs/graphql';
 import { GraphQLResolveInfo } from 'graphql';
-import { getSelect } from '~/util/select';
+import { makeGetSelect } from '~/util/select';
 import {
   WalletId,
   SetWalletNameArgs,
-  SetQuorumsArgs as SetWalletQuorumsArgs,
   WalletArgs,
+  UpsertWalletArgs,
+  RemoveWalletArgs,
 } from './wallets.args';
 import {
   connectOrCreateAccount,
@@ -26,6 +27,17 @@ import { Address, getWalletId, hashQuorum, Quorum, toWalletRef } from 'lib';
 import { Prisma } from '@prisma/client';
 import { UserAddr } from '~/decorators/user.decorator';
 import { SubgraphService } from '../subgraph/subgraph.service';
+import { UserInputError } from 'apollo-server-core';
+import { ProposableState, ProposableStatus } from './proposable.args';
+
+const getSelect = makeGetSelect<{
+  Wallet: Prisma.WalletSelect;
+}>({
+  Wallet: {
+    accountId: true,
+    ref: true,
+  },
+});
 
 @Resolver(() => Wallet)
 export class WalletsResolver {
@@ -33,11 +45,6 @@ export class WalletsResolver {
     private prisma: PrismaService,
     private subgraph: SubgraphService,
   ) {}
-
-  @ResolveField(() => String)
-  id(@Parent() acc: Wallet): string {
-    return getWalletId(acc.accountId, toWalletRef(acc.ref));
-  }
 
   @Query(() => Wallet, { nullable: true })
   async wallet(
@@ -90,6 +97,66 @@ export class WalletsResolver {
     });
   }
 
+  @ResolveField(() => String)
+  id(@Parent() w: Wallet): string {
+    return getWalletId(w.accountId, toWalletRef(w.ref));
+  }
+
+  @ResolveField(() => ProposableState, { nullable: true })
+  async state(@Parent() w: Wallet): Promise<ProposableState | null> {
+    const { createProposal, removeProposal } =
+      await this.prisma.wallet.findUniqueOrThrow({
+        where: {
+          accountId_ref: {
+            accountId: w.accountId,
+            ref: w.ref,
+          },
+        },
+        select: {
+          createProposal: {
+            select: {
+              hash: true,
+              createdAt: true,
+              submissions: {
+                where: { finalized: true },
+              },
+            },
+          },
+          removeProposal: {
+            select: {
+              hash: true,
+              createdAt: true,
+              submissions: {
+                where: { finalized: true },
+              },
+            },
+          },
+        },
+      });
+
+    // remove | deleted if it has a removeProposal that is newer than the create proposal
+    if (
+      removeProposal &&
+      (!createProposal || removeProposal.createdAt > createProposal.createdAt)
+    ) {
+      if (removeProposal.submissions.length) return null;
+
+      return {
+        status: ProposableStatus.remove,
+        proposedModificationHash: removeProposal.hash,
+      };
+    }
+
+    // Otherwise, active if the create proposal has a finalized submission otherwise it is added
+    if (createProposal?.submissions.length)
+      return { status: ProposableStatus.active };
+
+    return {
+      status: ProposableStatus.add,
+      proposedModificationHash: createProposal?.hash,
+    };
+  }
+
   @Mutation(() => Wallet)
   async setWalletName(
     @Args() { id, name }: SetWalletNameArgs,
@@ -110,18 +177,20 @@ export class WalletsResolver {
   }
 
   @Mutation(() => Wallet, { nullable: true })
-  async setWalletQuroums(
-    @Args() { id, quorums, txHash }: SetWalletQuorumsArgs,
+  async upsertWallet(
+    @Args() { id, name, quorums, proposalHash }: UpsertWalletArgs,
     @Info() info: GraphQLResolveInfo,
   ): Promise<Wallet> {
-    const existingTxHash = (
-      await this.prisma.wallet.findUnique({
-        where: {
-          accountId_ref: id,
+    // Only a single proposal can be active at a time for any given wallet
+    // Deleting prior unfinalized proposals related to the wallet will cascade delete a proposed wallet and quorums
+    await this.prisma.tx.deleteMany({
+      where: {
+        ...this.getActiveModificationProposalsWhere(id.accountId, id.ref),
+        hash: {
+          not: proposalHash,
         },
-        select: { txHash: true },
-      })
-    )?.txHash;
+      },
+    });
 
     return this.prisma.wallet.upsert({
       where: {
@@ -130,18 +199,27 @@ export class WalletsResolver {
       create: {
         account: connectOrCreateAccount(id.accountId),
         ref: id.ref,
-        txHash,
+        createProposal: {
+          connect: {
+            accountId_hash: {
+              accountId: id.accountId,
+              hash: proposalHash,
+            },
+          },
+        },
+        name,
         quorums: {
           create: quorums.map((quorum) =>
-            this.createQuorum(id, txHash, quorum),
+            this.createQuorum(id, proposalHash, quorum),
           ),
         },
       },
       update: {
-        ...(!existingTxHash && {
-          txHash: { set: { txHash } },
+        ...(name && {
+          name: { set: name },
         }),
         quorums: {
+          // Mark new quorums as being added by the proposal
           connectOrCreate: quorums.map((quorum) => {
             const hash = hashQuorum(quorum);
 
@@ -153,34 +231,88 @@ export class WalletsResolver {
                   hash,
                 },
               },
-              create: this.createQuorum(id, txHash, quorum),
+              create: this.createQuorum(id, proposalHash, quorum),
             };
           }),
-        },
+          // Mark quorums not included as being deleted by the proposal
+          // These quorums are all finalized; unfinalized modification proposals, and thus their quorums, were deleted above
+          updateMany: {
+            where: {
+              hash: {
+                notIn: quorums.map((quorum) => hashQuorum(quorum)),
+              },
+            },
+            data: {
+              removeProposalAccountId: id.accountId,
+              removeProposalHash: proposalHash,
+            },
+          },
+        } as Prisma.QuorumUpdateManyWithoutWalletNestedInput,
       },
       ...getSelect(info),
     });
   }
 
   @Mutation(() => Boolean)
-  deleteWallet(@Args() { id }: WalletArgs): boolean {
-    this.prisma.wallet.delete({
-      where: {
-        accountId_ref: id,
-      },
-    });
+  async deleteWallet(
+    @Args() { id, proposalHash }: RemoveWalletArgs,
+  ): Promise<boolean> {
+    const isActive = (
+      await this.prisma.wallet.findUniqueOrThrow({
+        where: {
+          accountId_ref: id,
+        },
+        select: {
+          createProposal: {
+            select: {
+              submissions: {
+                where: {
+                  finalized: true,
+                },
+                select: {
+                  txHash: true,
+                },
+              },
+            },
+          },
+        },
+      })
+    ).createProposal?.submissions.length;
+
+    // Propose deletion if it's active, otherwise delete
+    if (isActive) {
+      // Active; mark delete as occuring upon tx execution
+      if (!proposalHash)
+        throw new UserInputError("Can't delete active wallet without txHash");
+
+      this.prisma.wallet.update({
+        where: {
+          accountId_ref: id,
+        },
+        data: {
+          removeProposalHash: { set: proposalHash },
+        },
+      });
+    } else {
+      // Proposed; delete
+      this.prisma.wallet.delete({
+        where: {
+          accountId_ref: id,
+        },
+      });
+    }
 
     return true;
   }
 
   private createQuorum(
     id: WalletId,
-    txHash: string,
+    createProposalHash: string,
     quorum: Quorum,
   ): Prisma.QuorumUncheckedCreateWithoutWalletInput {
     return {
       hash: hashQuorum(quorum),
-      txHash,
+      createProposalHash,
       approvers: {
         create: quorum.map((approver) => ({
           account: { connect: { id: id.accountId } },
@@ -188,6 +320,47 @@ export class WalletsResolver {
           user: connectOrCreateUser(approver),
         })),
       },
+    };
+  }
+
+  private getActiveModificationProposalsWhere(
+    accountId: string,
+    walletRef: string,
+  ): Prisma.TxWhereInput {
+    return {
+      accountId: accountId,
+      walletRef,
+      submissions: {
+        none: {
+          finalized: true,
+        },
+      },
+      OR: [
+        {
+          proposedCreateWallet: {
+            ref: walletRef,
+          },
+        },
+        {
+          proposedRemoveWallet: {
+            ref: walletRef,
+          },
+        },
+        {
+          proposedCreateQuroums: {
+            some: {
+              walletRef,
+            },
+          },
+        },
+        {
+          proposedRemoveQuroums: {
+            some: {
+              walletRef,
+            },
+          },
+        },
+      ],
     };
   }
 }
